ooscalar

+++ src/main/scala/io/turntabl/grade/Grade.scala
@@ -1,3 +1,3 @@
 package io.turntabl.grade
 
-case class Grade(subject:String,grade: Int)
+case class Grade(subject:String,grade: Double)
